[Discover] Test for unique ids
